backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg[binary]>=3.1.0
python-multipart>=0.0.6
